pyproject.toml
[project]
name = "private"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-api-python-client>=2.170.0",
    "google-auth>=2.40.2",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]
